light on sphere

+++ includes/sphere.h
@@ -5,12 +5,12 @@ typedef struct	s_sp
 {
 	t_vec		*center;
 	float		radius;
-	int			color;
+	float		color;
 	struct s_sp	*next;
 }				t_sp;
 
-t_sp 	*sp_default(t_vec *center, float diameter, int color);
-void 	sp_list(t_sp **sp_list, t_vec *center, float diameter, int color);
+t_sp 	*sp_default(t_vec *center, float diameter, float color);
+void 	sp_list(t_sp **sp_list, t_vec *center, float diameter, float color);
 float	sp_crossing(t_cam *cam, t_vec *ray, t_sp *sp);
 
 #endif
